Fix Kurucz imported A values (100x too high from missing [Angstrom/nm]^2) (#14)
Bug found by @fmcneill49 resulting in A values for imported Kurucz data
(e.g. Sr, Y, Zr in Shingles et al. 2023) that were too high by 100x. So
far this appears to not alter the conclusion of Shingles+23:
 - the light curve is similar within 25%
- a similar spectral sequence is still seen, but evolving at a (~30%)
slower rate. However, it was already the case that the rate of time
evolution already did not AT2017gfo.
- smaller changes in the spectra features look like they will improve
the match to AT2017gfo. In Shingles+23, the spectra were heavily
smoothed, which obscured differences at this level of significance.

Fixed simulations for (unpublished) higher mass models are in-progress,
but will likely result is fairly large differences. This could help to
fix the excess opacity at early times, and lack of blue flux in the
spectra.

Co-authored-by: fmcneill49 <142821158+fmcneill49@users.noreply.github.com>